feat(backend): Implement image storage with Cloudinary, update data seeding files, add Elasticsearch support for subcategory

+++ backend/src/main/java/com/tuankhoi/backend/dto/response/APIResponse.java
@@ -19,5 +19,5 @@ public class APIResponse <T>{
 
     T result;
 
-    Long totalRecords;
+    Integer totalRecords;
 }
